feat: 이미지 등록 테스트

+++ src/main/java/com/january/guestbook/dto/MovieListDTO.java
@@ -10,7 +10,7 @@ public class MovieListDTO {
     private String title;
     private Long imageNumber;
     private int reviewCount;
-    private long gradeAverage;
+    private double gradeAverage;
 
     private LocalDateTime regDate;
 }
